+++ src/test/java/com/example/application/ApplicationIntegrationTest.java
@@ -1,5 +1,6 @@
 package com.example.application;
 
+import com.example.application.common.AbstractIntegrationTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.modulith.core.ApplicationModules;
 import org.springframework.modulith.docs.Documenter;
