+++ src/test/java/com/example/application/SchemeIntTest.java
@@ -2,6 +2,7 @@ package com.example.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.example.application.common.AbstractIntegrationTest;
 import net.ttddyy.dsproxy.support.ProxyDataSource;
 import org.junit.jupiter.api.Test;
